Revert "generate controller items"
This reverts commit fb0be9e6f2cce1ed72aed3f052aafda9ca9e207a.

+++ db/migrate/20220421060349_create_order_details.rb
@@ -2,7 +2,7 @@ class CreateOrderDetails < ActiveRecord::Migration[7.0]
   def change
     create_table :order_details do |t|
       t.integer :quantity
-      t.integer :item_price
+      t.float :item_price
 
       t.timestamps
     end
